requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch==2.0.0
torchvision==0.15.0
matplotlib>=3.4.3
seaborn>=0.11.2
pytest>=6.2.5
python-dotenv>=0.19.0
mlflow>=1.20.2
pyyaml>=5.4.1
joblib>=1.0.1 